Consolidate overlapping sema tests
