--- tests/test_sema_resolved_explicit_template_direct_call_ret0.cpp
@@ -1,19 +1,0 @@
-template <typename T>
-int addTwo(T value) {
-	return static_cast<int>(value) + 2;
-}
-
-namespace math {
-	template <typename T>
-	int multiplyBySix(T value) {
-		return static_cast<int>(value) * 6;
-	}
-}
-
-int main() {
-	if (addTwo<int>(40) != 42)
-		return 1;
-	if (math::multiplyBySix<int>(7) != 42)
-		return 2;
-	return 0;
-}
