--- tests/test_sema_resolved_inherited_member_call_ret0.cpp
@@ -1,19 +1,0 @@
-struct Base {
-	int base;
-
-	int addToBase(int value) {
-		return base + value;
-	}
-};
-
-struct Derived : Base {
-	int compute() {
-		base = 40;
-		return addToBase(2);
-	}
-};
-
-int main() {
-	Derived d;
-	return d.compute() - 42;
-}
