--- tests/test_sema_resolved_qualified_enum_constant_ret42.cpp
@@ -1,7 +1,0 @@
-enum class Numbers {
-	FortyTwo = 42,
-};
-
-int main() {
-	return static_cast<int>(Numbers::FortyTwo);
-}
