--- tests/test_sema_resolved_qualified_static_member_value_ret42.cpp
@@ -1,11 +1,0 @@
-template <typename T>
-struct Counter {
-	static int value;
-};
-
-template <typename T>
-int Counter<T>::value = static_cast<int>(sizeof(T)) + 38;
-
-int main() {
-	return Counter<int>::value;
-}
